+++ src/EventCentric/EventCentric/Repository/Mapping/SubscriptionEntity.cs
@@ -8,9 +8,9 @@ namespace EventCentric.Repository.Mapping
         public string StreamType { get; set; }
         public string Url { get; set; }
         public string Token { get; set; }
-        public int ProcessorBufferVersion { get; set; }
+        public long ProcessorBufferVersion { get; set; }
         public bool IsPoisoned { get; set; }
-        public int? PoisonEventCollectionVersion { get; set; }
+        public long? PoisonEventCollectionVersion { get; set; }
         public string DeadLetterPayload { get; set; }
         public string ExceptionMessage { get; set; }
         public DateTime CreationDate { get; set; }
