+++ src/EventCentric/EventCentric/Transport/NewRawEvent.cs
@@ -2,13 +2,13 @@
 {
     public class NewRawEvent
     {
-        public NewRawEvent(int eventCollectionVersion, string payload)
+        public NewRawEvent(long eventCollectionVersion, string payload)
         {
             this.EventCollectionVersion = eventCollectionVersion;
             this.Payload = payload;
         }
 
-        public int EventCollectionVersion { get; set; }
+        public long EventCollectionVersion { get; set; }
         public string Payload { get; private set; }
     }
 }
